+++ NEWS.md
@@ -1,6 +1,6 @@
 # data.io 1.3.0
 
-- `as.dataframe(table_object)` was broekn with 'tibble' 3.0.0. Now we use a different code to convert `table` objects into `tibble`/`dataframe`. 
+- `as.dataframe(table_object)` was broken with 'tibble' 3.0.0. Now we use a different code to convert `table` objects into `tibble`/`dataframe`. 
 
 - Rework of sources and 'pkgdown' web site added.